chore(quality): enforced ruff + mypy gates, and the defects they surfaced
Static analysis had never run here. Most of what it found was annotation debt,
but three items were real:

- Four `zip()` calls without `strict=`, two of them on paths where a length
  mismatch would silently produce a WRONG NUMBER rather than an error:
  `diagnostics.py` zips feature names to importances (a preprocessing change
  that adds or drops a column would misattribute every importance), and
  `serve.py` zips request rows to probabilities (a short result would silently
  skip logging rows). Now strict, with the reasoning at the call site.
- `registry.py` computed `incumbent_metric - min_delta` where the checker could
  not see that the branch guarantees non-None. The invariant held, but it lived
  only in the reader's head and this is the promotion gate. Now an explicit
  PromotionError, so if the metric read is ever made conditional the gate
  raises instead of computing `None - float`.
- `cli.py` reused `result`/`report` for four different types across mutually
  exclusive branches. Harmless at runtime, unreadable, and it hid the above.

The gate: ruff + mypy in CI as a separate `lint` job, curated rule set the repo
actually passes, E501 off with the reasoning recorded in pyproject (no formatter
is enforced — reformatting would destroy `git blame` on a commit log that is a
working engineering record).

B008 is scoped off for serve.py only: FastAPI's dependency-injection idiom IS a
function call in an argument default.

205 tests pass. ruff and mypy are both clean.

+++ src/pdm_mlops/detect_score.py
@@ -83,7 +83,7 @@ def _alarm_set(scores: np.ndarray, budget: float) -> np.ndarray:
     flags 0.6 % of rows is credited with exactly those, not the whole dataset.
     """
     n = len(scores)
-    k = max(1, int(round(budget * n)))
+    k = max(1, round(budget * n))
     kth = np.partition(scores, n - k)[n - k]  # value at the budget boundary
     if (scores == kth).sum() > k - (scores > kth).sum():
         # ties overflow the budget at the boundary → only strictly-greater rows alarm
